@@ -3,58 +3,85 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "postinstall": "npx husky install",
+    "prepare": "npx simple-git-hooks",
     "ng": "ng",
-    "start": "ng serve",
+    "dev": "ng serve",
+    "dev:host": "ng serve --host 0.0.0.0 --disable-host-check",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "test:run": "ng test --watch=false --browsers ChromeHeadless",
-    "lint": "ng lint"
+    "test:headless": "ng test --watch=false --browsers ChromeHeadless",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "lint:staged": "pnpm lint-staged",
+    "stylelint": "stylelint **/*.{css,pcss,less,scss,sass} --quiet-deprecation-warnings",
+    "stylelint:fix": "stylelint **/*.{css,pcss,less,scss,sass} --quiet-deprecation-warnings --fix",
+    "format": "prettier --write ."
+  },
+  "simple-git-hooks": {
+    "pre-commit": "pnpm lint-staged",
+    "pre-push": "pnpm build"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": [
+      "pnpm lint:fix",
+      "pnpm prettier"
+    ],
+    "*.{css,pcss,less,scss,sass}": [
+      "pnpm stylelint:fix",
+      "pnpm prettier"
+    ]
   },
   "dependencies": {
-    "@angular/animations": "~14.0.1",
-    "@angular/common": "~14.0.1",
-    "@angular/compiler": "~14.0.1",
-    "@angular/core": "~14.0.1",
-    "@angular/forms": "~14.0.1",
-    "@angular/platform-browser": "~14.0.1",
-    "@angular/platform-browser-dynamic": "~14.0.1",
-    "@angular/router": "~14.0.1",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.4.0",
-    "zone.js": "~0.11.5"
+    "@angular/animations": "^16.0.0",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.0.1",
-    "@angular-eslint/builder": "13.4.0",
-    "@angular-eslint/eslint-plugin": "13.4.0",
-    "@angular-eslint/eslint-plugin-template": "13.4.0",
-    "@angular-eslint/schematics": "13.4.0",
-    "@angular-eslint/template-parser": "13.4.0",
-    "@angular/cli": "14.0.1",
-    "@angular/compiler-cli": "~14.0.1",
-    "@types/jasmine": "~4.0.0",
-    "jasmine-core": "~4.1.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "@angular-devkit/build-angular": "^16.0.1",
+    "@angular-eslint/builder": "16.0.1",
+    "@angular-eslint/eslint-plugin": "16.0.1",
+    "@angular-eslint/eslint-plugin-template": "16.0.1",
+    "@angular-eslint/schematics": "16.0.1",
+    "@angular-eslint/template-parser": "16.0.1",
+    "@angular/cli": "~16.0.1",
+    "@angular/compiler-cli": "^16.0.0",
+    "@culur/stylelint-config": "^1.2.0",
+    "@iconify/json": "^2.2.64",
+    "@iconify/tailwind": "^0.1.2",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/forms": "^0.5.3",
+    "@tailwindcss/typography": "^0.5.9",
+    "@types/jasmine": "~4.3.0",
+    "@types/node": "^20.1.3",
+    "@typescript-eslint/eslint-plugin": "5.59.2",
+    "@typescript-eslint/parser": "5.59.2",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "eslint": "^8.17.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-angular": "4.1.0",
-    "eslint-plugin-import": "2.25.2",
-    "eslint-plugin-jsdoc": "36.1.0",
-    "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-prettier": "4.0.0",
-    "husky": "7.0.2",
-    "postcss": "^8.4.14",
-    "prettier": "2.4.1",
-    "prettier-eslint": "13.0.0",
-    "tailwindcss": "3.1.2",
-    "typescript": "4.7.3"
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "lint-staged": "^13.2.2",
+    "postcss": "^8.4.23",
+    "prettier": "^2.8.8",
+    "prettier-eslint": "^15.0.1",
+    "prettier-plugin-tailwindcss": "^0.2.8",
+    "simple-git-hooks": "^2.8.1",
+    "stylelint": "^15.6.1",
+    "tailwindcss": "^3.3.2",
+    "typescript": "~5.0.2"
   }
 }
