@@ -1,26 +1,22 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, inject, OnInit } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { AuthService } from '@lib/services';
 import { ThemeService } from '@lib/services/theme';
-import { Observable } from 'rxjs';
 import { LayoutHorizontalComponent } from './lib/components/layouts/layout-horizontal/layout-horizontal.component';
 
 @Component({
-  selector: 'app-root',
-  standalone: true,
-  imports: [CommonModule, RouterModule, LayoutHorizontalComponent],
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+    selector: 'app-root',
+    standalone: true,
+    imports: [CommonModule, RouterModule, LayoutHorizontalComponent],
+    templateUrl: './app.component.html',
 })
 export class AppComponent implements OnInit {
-  isLoggedIn$!: Observable<boolean>;
+    isAuthenticated$ = inject(AuthService).isAuthenticated$;
 
-  constructor(private _authService: AuthService, private _themeService: ThemeService) {}
+    private readonly _themeService = inject(ThemeService);
 
-  ngOnInit(): void {
-    this.isLoggedIn$ = this._authService.isLoggedIn$;
-
-    this._themeService.init();
-  }
+    ngOnInit(): void {
+        this._themeService.init();
+    }
 }
