@@ -1,18 +1,18 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { REPOSITORY_URL } from '@lib/constants';
+import { PACKAGE_JSON, providePackageJson } from '@lib/providers';
 import { LogoComponent } from '../logo/logo.component';
 
 @Component({
-  selector: 'app-footer',
-  standalone: true,
-  imports: [CommonModule, RouterModule, LogoComponent],
-  templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    selector: 'app-footer',
+    standalone: true,
+    imports: [CommonModule, RouterModule, LogoComponent],
+    providers: [providePackageJson()],
+    templateUrl: './footer.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class FooterComponent {
-  readonly repositoryURL = REPOSITORY_URL;
-  readonly currentYear = new Date().getFullYear();
+    readonly packageJson = inject(PACKAGE_JSON);
+    readonly currentYear = new Date().getFullYear();
 }
