@@ -3,11 +3,10 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { FooterComponent, NavbarComponent } from '@lib/components';
 
 @Component({
-  selector: 'app-layout-horizontal',
-  standalone: true,
-  imports: [CommonModule, NavbarComponent, FooterComponent],
-  templateUrl: './layout-horizontal.component.html',
-  styleUrls: ['./layout-horizontal.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    selector: 'app-layout-horizontal',
+    standalone: true,
+    imports: [CommonModule, NavbarComponent, FooterComponent],
+    templateUrl: './layout-horizontal.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LayoutHorizontalComponent {}
