@@ -3,11 +3,10 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 @Component({
-  selector: 'app-logo',
-  standalone: true,
-  imports: [CommonModule, RouterModule],
-  templateUrl: './logo.component.html',
-  styleUrls: ['./logo.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush,
+    selector: 'app-logo',
+    standalone: true,
+    imports: [CommonModule, RouterModule],
+    templateUrl: './logo.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LogoComponent {}
