@@ -1,5 +1,5 @@
 export enum UserRole {
-  Admin = 'ADMIN',
-  User = 'USER',
-  Guest = 'GUEST',
+    Admin = 'ADMIN',
+    User = 'USER',
+    Guest = 'GUEST',
 }
