@@ -1,2 +1,1 @@
-export { AuthGuard } from './auth.guard';
-export { NoAuthGuard } from './no-auth.guard';
+export * from './auth.guard';
