@@ -1,31 +1,31 @@
-import {
-  HttpEvent,
-  HttpHandler,
-  HttpInterceptor,
-  HttpRequest,
-} from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { HttpInterceptorFn } from '@angular/common/http';
+import { inject } from '@angular/core';
 import { environment } from '@env/environment';
-import { Observable } from 'rxjs';
+import { AuthService } from '@lib/services';
 
-@Injectable()
-export class JwtInterceptor implements HttpInterceptor {
-  intercept(
-    request: HttpRequest<unknown>,
-    next: HttpHandler,
-  ): Observable<HttpEvent<unknown>> {
-    const isLoggedIn = true;
-    const token = 'ABC';
-    const isApiUrl = request.url.startsWith(environment.apiUrl);
+/**
+ * Interceptor that adds an Authorization header to requests that are authenticated and target the API URL.
+ *
+ * @param request The request object.
+ * @param next The next interceptor in the chain.
+ *
+ * @returns The next Observable.
+ */
+export const jwtInterceptor: HttpInterceptorFn = (request, next) => {
+    const authService = inject(AuthService);
 
-    if (isLoggedIn && isApiUrl) {
-      request = request.clone({
-        setHeaders: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
+    const isRequestAuthorized = authService.isAuthenticated && request.url.startsWith(environment.apiUrl);
+
+    if (isRequestAuthorized) {
+        const clonedRequest = request.clone({
+            setHeaders: {
+                // eslint-disable-next-line @typescript-eslint/naming-convention
+                Authorization: `Bearer ${'JWT TOKEN'}`,
+            },
+        });
+
+        return next(clonedRequest);
     }
 
-    return next.handle(request);
-  }
-}
+    return next(request);
+};
