@@ -3,14 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { AuthService } from './auth.service';
 
 describe('AuthService', () => {
-  let service: AuthService;
+    let service: AuthService;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({ teardown: { destroyAfterEach: false } });
-    service = TestBed.inject(AuthService);
-  });
+    beforeEach(() => {
+        TestBed.configureTestingModule({ teardown: { destroyAfterEach: false } });
+        service = TestBed.inject(AuthService);
+    });
 
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
 });
