@@ -3,22 +3,22 @@ import { storage } from '@lib/utils/storage/storage.utils';
 import { BehaviorSubject } from 'rxjs';
 
 @Injectable({
-  providedIn: 'root',
+    providedIn: 'root',
 })
 export class AuthService {
-  isLoggedIn$ = new BehaviorSubject<boolean>(!!storage.getItem('App/session'));
+    isAuthenticated$ = new BehaviorSubject<boolean>(!!storage.getItem('appSession'));
 
-  get isLoggedIn(): boolean {
-    return this.isLoggedIn$.getValue();
-  }
+    get isAuthenticated(): boolean {
+        return this.isAuthenticated$.getValue();
+    }
 
-  login(): void {
-    storage.setItem('App/session', { user: 'some-user-id', token: 'abc' });
-    this.isLoggedIn$.next(true);
-  }
+    login(): void {
+        storage.setItem('appSession', { user: 'some-user-id', token: 'abc' });
+        this.isAuthenticated$.next(true);
+    }
 
-  logout(): void {
-    storage.removeItem('App/session');
-    this.isLoggedIn$.next(false);
-  }
+    logout(): void {
+        storage.removeItem('appSession');
+        this.isAuthenticated$.next(false);
+    }
 }
