@@ -1,28 +1,7 @@
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { enableProdMode, importProvidersFrom } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
-import { RouterModule } from '@angular/router';
 import { AppComponent } from './app/app.component';
-import { routes } from './app/app.routes';
-import { JwtInterceptor, ServerErrorInterceptor } from './app/lib/interceptors';
-import { environment } from './environments/environment';
+import { appConfig } from './app/app.config';
 
-if (environment.production) {
-  enableProdMode();
-}
-
-bootstrapApplication(AppComponent, {
-  providers: [
-    importProvidersFrom(RouterModule.forRoot(routes), HttpClientModule),
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: ServerErrorInterceptor,
-      multi: true,
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: JwtInterceptor,
-      multi: true,
-    },
-  ],
-}).catch((error) => console.error(error));
+bootstrapApplication(AppComponent, appConfig).catch((error) => {
+    console.error(error);
+});
